fixes to sub
